Convert TooltipButton to TypeScript
Refactored; TypeScript; functional component.
